change publish script

+++ publish.sh
@@ -3,10 +3,8 @@
 WORKING_DIR=_site
 PUBLISH_DIR=docs
 if [ -d "$WORKING_DIR" ]; then rm -Rf $WORKING_DIR; fi
-if [ -d "$PUBLISH_DIR" ]; then
-  rm -Rf $PUBLISH_DIR
-  mkdir $PUBLISH_DIR
-fi
+if [ -d "$PUBLISH_DIR" ]; then rm -Rf $PUBLISH_DIR; fi
+mkdir $PUBLISH_DIR
 JEKYLL_ENV=production jekyll build
 cp -r $WORKING_DIR/** $PUBLISH_DIR
 git add $PUBLISH_DIR
